requirements.txt
# Django y dependencias principales
Django==5.2.6
djangorestframework==3.16.1

# Filtros y paginación
django-filter==25.1

# Tareas programadas
django-crontab==0.7.1

# HTTP requests
requests==2.32.5

# Notificaciones
python-telegram-bot==22.5

# Para desarrollo y testing (opcional)
coverage==7.6.9
pytest-django==4.8.0

# Para producción (opcional)
gunicorn==21.2.0
psycopg2-binary==2.9.7  # PostgreSQL
redis==4.6.0  # Cache
django-redis==5.3.0  # Django Redis integration
sentry-sdk==1.32.0  # Error monitoring

# Para API documentation (opcional)
drf-yasg==1.21.7  # Swagger/OpenAPI

# Para rate limiting (opcional)
django-ratelimit==4.1.0